requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.3
pymysql==1.0.3
discord.py==2.6.1
python-dotenv==1.0.0
SQLAlchemy==2.0.19
